sprint 5 2/3: transfer Services and Models from MovieQuizViewController
