+++ MovieQuiz/Helpers/Array+Extensions.swift
@@ -5,3 +5,14 @@ extension Array {
         indices ~= index ? self[index] : nil
     }
 }
+
+/*
+ subscript(index: Int) -> Int {
+     get {
+         // Возвращаем соответствующее значение
+     }
+     set(newValue) {
+         // Устанавливаем подходящее значение
+     }
+ }
+ */
