+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Anton Silenin on 08.04.2025.
+//
+
+import Foundation
+import UIKit
+
+struct QuizResultsViewModel {
+  let title: String
+  let text: String
+  let buttonText: String
+}
