+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -10,78 +10,6 @@ final class MovieQuizViewController: UIViewController {
         show(quiz: viewModel)
     }
     
-    struct QuizQuestion {
-      let image: String
-      let text: String
-      let correctAnswer: Bool
-    }
-    
-    struct QuizStepViewModel {
-      let image: UIImage
-      let question: String
-      let questionNumber: String
-    }
-    
-    struct QuizResultsViewModel {
-      let title: String
-      let text: String
-      let buttonText: String
-    }
-
-    private let questions: [QuizQuestion]  =    [
-        QuizQuestion(
-             image: "The Godfather",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: true
-        ),
-        QuizQuestion(
-             image: "The Dark Knight",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: true
-        ),
-        QuizQuestion(
-             image: "Kill Bill",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: true
-        ),
-        QuizQuestion(
-             image: "The Avengers",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: true
-        ),
-        QuizQuestion(
-             image: "Deadpool",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: true
-        ),
-        QuizQuestion(
-             image: "The Green Knight",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: true
-        ),
-        QuizQuestion(
-             image: "Old",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: false
-        ),
-        QuizQuestion(
-             image: "The Ice Age Adventures of Buck Wild",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: false
-        ),
-        QuizQuestion(
-             image: "Tesla",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: false
-        ),
-        QuizQuestion(
-             image: "Vivarium",
-             text: "Рейтинг этого фильма больше чем 6?",
-             correctAnswer: false
-        )
-     
-    ]
-    
     private var currentQuestionIndex = 0
     
     private var correctAnswers = 0
@@ -104,7 +32,9 @@ final class MovieQuizViewController: UIViewController {
     private func showAnswerResult(isCorrect: Bool) {
         if isCorrect { correctAnswers += 1}
         self.showBorder(isCorrect: isCorrect)
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
+        
+        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) { [weak self] in
+            guard let self = self else { return }
             self.hideBoarder()
             self.showNextQuestionOrResults()
         }
@@ -141,13 +71,18 @@ final class MovieQuizViewController: UIViewController {
             message: result.text,
             preferredStyle: .alert)
         
-        let action = UIAlertAction(title:result.buttonText, style: .default) { _ in
+        let action = UIAlertAction(title: result.buttonText, style: .default) { [weak self] _ in
+            guard let self = self else { return }
+            
             self.currentQuestionIndex = 0
             self.correctAnswers = 0
+            
             let firstQuestion = self.questions[self.currentQuestionIndex]
             let viewModel = self.convert(model: firstQuestion)
             self.show(quiz: viewModel)
         }
+        
+        
         alert.addAction(action)
         self.present(alert, animated: true, completion: nil)
     }
